# docs/swagger.yaml
basePath: /api
host: 114.132.40.112:9502
info:
  contact:
    email: pltrueover@gmail.com
    name: go-core
    url: https://im.pltrue.top
  description: go-im 服务端
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: go-im  接口文档
  version: "2.0"
paths:
...
  /ReadMessage:
    get:
      description: 历史消息读取
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 图片上传
        in: formData
        name: voice
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 历史消息读取[废弃]
      tags:
      - 历史消息读取
...
  /Update:
    put:
      description: 更新用户数据
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 个性签名
        in: formData
        name: bio
        type: string
      - description: 性别
        in: formData
        name: six
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 更新用户数据
      tags:
      - 更新用户数据
...
  /UpdatePwd:
    post:
      consumes:
      - multipart/form-data
      description: 更新密码
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 旧密码
        in: formData
        name: password
        required: true
        type: string
      - description: 新密码
        in: formData
        name: new_password
        required: true
        type: string
      - description: 重复密码
        in: formData
        name: password_confirm
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 更新密码
      tags:
      - 更新密码
...
  /UsersList:
    get:
      description: 根据昵称查询非好友用户列表
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 账号
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 根据昵称查询非好友用户列表
      tags:
      - 根据昵称查询非好友用户列表
  /bindingEmail:
    post:
      consumes:
      - multipart/form-data
      description: 绑定用户邮箱接口
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 绑定用户邮箱
      tags:
      - 绑定用户邮箱
  /login:
    post:
      consumes:
      - multipart/form-data
      description: 登录接口
      parameters:
      - description: 账号
        in: formData
        name: name
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 客户端类型 0.网页端登录 1.设备端登录
        in: formData
        name: client_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 这是一个登录接口
      tags:
      - 登录接口
  /me:
    post:
      description: 获取用户信息接口
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 获取用户信息接口
      tags:
      - 获取用户信息接口
  /registered:
    post:
      consumes:
      - multipart/form-data
      description: 注册用户接口
      parameters:
      - description: 用户名
        in: formData
        name: name
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 确认密码
        in: formData
        name: password_confirm
        required: true
        type: string
      - description: 验证码
        in: formData
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 注册用户
      tags:
      - 注册用户
  /seedRegisteredEmail:
    get:
      description: 发送注册邮箱验证码接口
      parameters:
      - description: 邮箱
        in: query
        name: email
        type: string
      responses:
        "200":
          description: ""
      summary: 发送注册邮箱验证码
      tags:
      - 发送注册邮箱验证码接口
  /userDetails:
    get:
      consumes:
      - multipart/form-data
      description: 根据id获取用户详情
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 邮箱
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 根据id获取用户详情
      tags:
      - 根据id获取用户详情
swagger: "2.0"